requirements.txt
matplotlib==3.9.2
pandas==2.2.2
python_dateutil==2.9.0.post0
yfinance==0.2.41
